Display search context in the dropdown menu
